+++ allcommands.ps1
@@ -19,23 +19,40 @@ function Get-MAML
     $Module,
 
     # The CommandInfo object (returned from Get-Command).
-    [Parameter(Mandatory=$true,ParameterSetName='FromCommandInfo', ValueFromPipeline=$true)]
+    [Parameter(Mandatory=$true,ParameterSetName='FromCommandInfo', ValueFromPipeline=$true)]    
     [Management.Automation.CommandInfo[]]
     $CommandInfo,
 
     # If set, the generated MAML will be compact (no extra whitespace or indentation).  If not set, the MAML will be indented.
+    [Parameter(ValueFromPipelineByPropertyName)]
     [switch]
     $Compact,
     
     # If set, will return the MAML as an XmlDocument.  The default is to return the MAML as a string.
+    [Parameter(ValueFromPipelineByPropertyName)]
     [switch]
     $XML,
     
     # If set, the generate MAML will not contain a version number.  
     # This slightly reduces the size of the MAML file, and reduces the rate of changes in the MAML file.
+    [Parameter(ValueFromPipelineByPropertyName)]
     [Alias('Unversioned')]
     [switch]
-    $NoVersion)
+    $NoVersion,
+    
+    # A list of command types to skip.
+    # If not provided, all types of commands from the module will be saved as a markdown document.
+    [Parameter(ValueFromPipelineByPropertyName)]
+    [Alias('SkipCommandTypes','ExcludeCommandType','ExcludeCommandTypes')]
+    [Management.Automation.CommandTypes[]]
+    $SkipCommandType,
+
+    # If set, will include aliases in the MAML output.
+    [Parameter(ValueFromPipelineByPropertyName)]
+    [Alias('IncludeAliases')]
+    [switch]
+    $IncludeAlias
+    )
     
     begin {
         # First, we need to create a list of all commands we encounter (so we can process them at the end)
@@ -297,7 +314,7 @@ function Get-MAML
     process {
         
         if ($PSCmdlet.ParameterSetName -eq 'ByName') { # If we're getting comamnds by name,
-            $CommandInfo = @(foreach ($n in $name) { 
+            $CommandInfo = @(foreach ($n in $name) {
                 $ExecutionContext.InvokeCommand.GetCommands($N,'Function,Cmdlet', $true) # find each command (treating Name like a wildcard).
             })
         }
@@ -311,8 +328,18 @@ function Get-MAML
 
 
         $filteredCmds = @(foreach ($ci in $CommandInfo) { # Filter the list of commands 
-            if ($ci -is [Management.Automation.AliasInfo] -or # (throw out aliases and applications).
-                $ci -is [Management.Automation.ApplicationInfo]) { continue }
+             # (throw out applications)
+            if ($ci -is [Management.Automation.ApplicationInfo]) {
+                continue
+            }
+
+            if ($ci -is [Management.Automation.AliasInfo]) {
+                if (-not $IncludeAlias) { continue }
+            }
+
+            if ($SkipCommandType -and 
+                $SkipCommandType -contains $ci.CommandType) { continue } # (throw out any we were told to skip)
+
             $ci 
         })
          
@@ -480,7 +507,7 @@ function Get-MarkdownHelp {
                         if ($Rename) {
                             $helpObj | Add-Member NoteProperty Rename $Rename -Force
                         } elseif ($IsHelpAboutAlias) {
-                            $helpObj | Add-Member NoteProperty Rename $gotHelp.Name -Force
+                            $helpObj | Add-Member NoteProperty Rename $getHelp.Name -Force
                         }
 
                         # * `-SectionOrder` will become `[string[]] .SectionOrder`
@@ -1310,6 +1337,19 @@ function Save-MAML
     [Alias('Unversioned')]
     [switch]
     $NoVersion,
+
+    # A list of command types to skip.
+    # If not provided, all types of commands from the module will be saved as a markdown document.
+    [Parameter(ValueFromPipelineByPropertyName)]
+    [Alias('SkipCommandTypes','ExcludeCommandType','ExcludeCommandTypes')]
+    [Management.Automation.CommandTypes[]]
+    $SkipCommandType,
+
+    # If set, will include aliases in the MAML output.
+    [Parameter(ValueFromPipelineByPropertyName)]
+    [Alias('IncludeAliases')]
+    [switch]
+    $IncludeAlias,
     
     # If set, will return the files that were generated.
     [switch]
@@ -1338,6 +1378,10 @@ function Save-MAML
             if (-not $getMAML.Parameters.ContainsKey($k)) { # that wasn't in Get-MAML.
                 $splat.Remove($k)
             }
+        }
+
+        if ($SkipCommandType -and -not $splat.SkipCommandType) {
+            $splat.SkipCommandType = $SkipCommandType
         }
 
         if (-not $Culture) { # If -Culture wasn't provided, use the current culture